+++ .mocharc.cjs
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  "node-option": ["import=tsx"],
+  // require: ['./test/setup.js'],
+  extension: ["ts"],
+  spec: ["test/**/*.spec.ts"]
+};